作者検索API実装
