+++ nest-backend/src/novels/dto/search-novels.dto.ts
@@ -1,3 +1,3 @@
 export class SearchNovelsDto {
-  keyword: string;
+  title: string;
 }
