+++ nest-backend/src/novels/novels.module.ts
@@ -17,8 +17,8 @@ import { SearchNovelsService } from './services/search-novels.service';
 // import { CommentsService } from '../comments/comments.service';
 import { GetCommentsByNovelIdService } from '../comments/service/get-comments-by-novelid.service';
 import { DeleteCommentsService } from '../comments/service/delete-comments.service';
-import { MongoSearchNovelRepository } from './repositories/mongo-search-novels.repositories';
-import { PostgresNovelRepository } from './repositories/postgres-search-novels.repositories';
+import { MongoSearchNovelRepository } from './repositories/search-novels/mongo-search-novels.repositories';
+import { PostgresNovelRepository } from './repositories/search-novels/postgres-search-novels.repositories';
 
 @Module({
   imports: [
